move utility func to store utility
