+++ src/sing/utility.ts
@@ -1,38 +1,3 @@
-import { CharacterInfo, StyleInfo, StyleType } from "@/type/preload";
-
-export const isSingingStyle = (styleInfo: StyleInfo) => {
-  return styleInfo.styleType === "humming" || styleInfo.styleType === "sing";
-};
-
-export const filterCharacterInfosByStyleType = (
-  characterInfos: CharacterInfo[],
-  styleType: StyleType | "singerLike"
-): CharacterInfo[] => {
-  const withStylesFiltered: CharacterInfo[] = characterInfos.map(
-    (characterInfo) => {
-      const styles = characterInfo.metas.styles.filter((styleInfo) => {
-        // singerLike：歌う系のスタイル
-        if (styleType === "singerLike") {
-          return isSingingStyle(styleInfo);
-        }
-        // talk：singerLike以外のスタイル。
-        // styleTypeが存在しない（マルチエンジン）場合があるので、「singerLike以外」をtalkとして扱っている。
-        if (styleType === "talk") {
-          return !isSingingStyle(styleInfo);
-        }
-        return styleInfo.styleType === styleType;
-      });
-      return { ...characterInfo, metas: { ...characterInfo.metas, styles } };
-    }
-  );
-
-  const withoutEmptyStyles = withStylesFiltered.filter(
-    (characterInfo) => characterInfo.metas.styles.length > 0
-  );
-
-  return withoutEmptyStyles;
-};
-
 export function round(value: number, digits: number) {
   const powerOf10 = 10 ** digits;
   return Math.round(value * powerOf10) / powerOf10;
